requirements.txt
langchain>=0.1.0
langchain-google-genai>=0.0.5
langchain-community>=0.0.10
google-generativeai>=0.3.2
chromadb>=0.4.22
streamlit>=1.29.0
python-dotenv>=1.0.0
tiktoken>=0.5.2
pypdf>=3.17.4
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
pandas>=2.1.4
numpy>=1.26.0
